+++ packages/offix-datastore/src/storage/api/StorageAdapter.ts
@@ -1,5 +1,4 @@
 import { PredicateFunction } from "../../predicates";
-import { CRUDEvents } from "./CRUDEvents";
 
 /**
  * This interface defines the API that is required
@@ -48,23 +47,3 @@ export interface StorageAdapter {
   remove(storeName: string, predicate?: PredicateFunction): Promise<any | any[]>;
 }
 
-/**
- * StoreChangeEvent is an event emitted whenever
- * a change has occurred on the local store
-*/
-export interface StoreChangeEvent {
-  /**
-   * The type of change event that just occurred
-   */
-  eventType: CRUDEvents;
-
-  /**
-   * The data that was affected by the change
-   */
-  data: any;
-
-  /**
-   * The name store that was changed
-   */
-  storeName: string;
-}
