--- packages/offix-datastore/src/storage/core.ts
@@ -1,5 +1,0 @@
-import { v1 as uuidv1 } from "uuid";
-
-export function generateId() {
-    return uuidv1();
-}
