+++ packages/offix-datastore/src/storage/index.ts
@@ -1,1 +1,5 @@
-export * from "./Storage";
+export * from "./LocalStorage";
+export * from "./api/CRUDEvents";
+export * from "./api/StorageAdapter";
+export * from "./api/StoreChangeEvent";
+
